moving files and sharing code
